fix: Revert "feat: improve cli detection" (#65)
Reverts localstack/localstack-toolkit-vscode#60

Step 2 in bringing main back to the state before
https://github.com/localstack/localstack-toolkit-vscode/pull/60 that
introduced a regression. Changes from #60 are to be reapplied in a
different manner to prevent the regression.

This bring `main` to a releaseable state again.

Note: step 1 is #64, had to be reverted as well because its changes were
made on top of #60 ones and cannot be isolated.